+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Find (and store) credentials by only a hostname as the machine in netrc
 * Make the command start up faster by lazy loading some gems
 
 # 4.18.1